Add --with-lg-tcache-limit to increase the max supported tcaches
In the Couchbase Server Data Service (KV Engine), we use one tcache per
arena, per thread (multipies).

We have one arena per bucket, and we currently support 30 buckets, but
might eventually bump this up to 100 buckets per instance.

However, because we allocate a tcache per thread as well, the number of
tcache which we use ends up being num-buckets x num-threads. Tcaches are
allocated when a thread decides to use a bucket's arena, allocating
tcaches lazily, so not all threads will need a tcache, but most will.

jemalloc can automatically allocate and make use of a tcache, but we've
found that can result in incorrect accounting of memory stats per-arena,
which is a deal-breaker for us, because we rely on some of these stats
to calculate per-arena memory fragmentation.

The old limit on the number of tcaches was largely dependant on the
representation of the tcache ID in the 32-bit flags through which it was
specified to je_mallocx.

---

Flag bits:
a: arena
t: tcache
.: configurable, dependent on --with-lg-tcache-limit:
   The higher this value, the more of the configurable bits will be teated as
   tcache bits (the rest are arena bits).
0: unused
z: zero
n: alignment

Old representation:
  aaaaaaaa aaaatttt tttttttt 0znnnnnn
New representation:
  aaaaaaa. ........ ....tttt 0znnnnnn

Before this change, we had 12 bits for the arena index, 12 bits for the
tcache ID, giving us 4094 (tcache 0 and 1 reserved) tcaches and 4096
arenas.

By using more bits from the arena index representation, we can increase
the representation of the tcache ID to be wider/narrower, allowing up
to 2^17 valid tcache IDs to be specified (but 0 and 1 are reserved by
jemalloc). In this example, however, the number of areans falls to 128.

(cherry picked from commit 89033437882ccc253b15a0e131594f0df0f8b58c)